Code cleanups on behalf of Lebedev, see issue 1332

+++ rtengine/diagonalcurves.cc
@@ -196,7 +196,7 @@ void DiagonalCurve::NURBS_set () {
         // TODO: Speeding-up the interface by caching the polyline, instead of rebuilding it at each action on sliders !!!
         nbr_points = (int)(((double)(ppn+N-2) * sc_length[i/3] )/ total_length);
         if (nbr_points<0){
-            for(int it=0;it < sc_x.size(); it+=3) printf("sc_length[%d/3]=%f \n",it,sc_length[it/3]);
+            for(size_t it=0;it < sc_x.size(); it+=3) printf("sc_length[%zu/3]=%f \n",it,sc_length[it/3]);
             printf("NURBS diagonal curve: error detected!\n i=%d nbr_points=%d ppn=%d N=%d sc_length[i/3]=%f total_length=%f",i,nbr_points,ppn,N,sc_length[i/3],total_length);
             exit(0);
         }
