+++ rtgui/toolpanelcoord.cc
@@ -176,7 +176,7 @@ ToolPanelCoordinator::ToolPanelCoordinator () : ipc(NULL)  {
     toolPanelNotebook->set_scrollable ();
     toolPanelNotebook->show_all ();
 
-    for (int i=0; i<toolPanels.size(); i++)
+    for (size_t i=0; i<toolPanels.size(); i++)
         toolPanels[i]->setListener (this);
 
     whitebalance->setWBProvider (this);
@@ -223,7 +223,7 @@ void ToolPanelCoordinator::panelChanged (rtengine::ProcEvent event, const Glib::
     int changeFlags=refreshmap[(int)event];
 
     ProcParams* params = ipc->beginUpdateParams ();
-    for (int i=0; i<toolPanels.size(); i++)
+    for (size_t i=0; i<toolPanels.size(); i++)
         toolPanels[i]->write (params);
 
     // Compensate rotation on flip
@@ -259,7 +259,7 @@ void ToolPanelCoordinator::panelChanged (rtengine::ProcEvent event, const Glib::
 
     hasChanged = true;
 
-    for (int i=0; i<paramcListeners.size(); i++)
+    for (size_t i=0; i<paramcListeners.size(); i++)
         paramcListeners[i]->procParamsChanged (params, event, descr);
 }
 
@@ -320,14 +320,14 @@ void ToolPanelCoordinator::profileChange  (const PartialProfile *nparams, rtengi
 
     hasChanged = event != rtengine::EvProfileChangeNotification;
 
-    for (int i=0; i<paramcListeners.size(); i++)
+    for (size_t i=0; i<paramcListeners.size(); i++)
         paramcListeners[i]->procParamsChanged (params, event, descr);
 }
 
 void ToolPanelCoordinator::setDefaults (ProcParams* defparams) {
 
     if (defparams)
-        for (int i=0; i<toolPanels.size(); i++) 
+	    for (size_t i=0; i<toolPanels.size(); i++)
             toolPanels[i]->setDefaults (defparams);
 }
 
@@ -370,7 +370,7 @@ void ToolPanelCoordinator::closeImage () {
 void ToolPanelCoordinator::readOptions () {
 
     crop->readOptions (); 
-    for (int i=0; i<options.tpOpen.size(); i++)
+    for (size_t i=0; i<options.tpOpen.size(); i++)
         if (i<expList.size())
             expList[i]->set_expanded (options.tpOpen[i]);
 }
@@ -379,7 +379,7 @@ void ToolPanelCoordinator::writeOptions () {
 
     crop->writeOptions (); 
     options.tpOpen.clear ();
-    for (int i=0; i<expList.size(); i++)
+    for (size_t i=0; i<expList.size(); i++)
         options.tpOpen.push_back (expList[i]->get_expanded ());
 }
 
@@ -519,7 +519,7 @@ void ToolPanelCoordinator::foldAllButOne (Gtk::Box* parent, FoldableToolPanel* o
 
 	FoldableToolPanel* currentTP;
 
-    for (int i=0; i<toolPanels.size(); i++) {
+	for (size_t i=0; i<toolPanels.size(); i++) {
 	currentTP = static_cast<FoldableToolPanel*>(toolPanels[i]);
         if (currentTP->getParent() == parent) {
             // Section in the same tab, we unfold it if it's not the one that has been clicked
